[FIX]ビルドが通らなかったのでエクスポートすることにより調整
