+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,54 +1,5 @@
 import NextAuth from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import type { NextAuthOptions } from "next-auth";
-import { prismaInitalied as prisma } from "@/src/lib/PrismaInitialized";
-import { getServerSession as originalGetServerSession } from "next-auth";
-import { cache } from "react";
-
-const authOptions: NextAuthOptions = {
-  secret: process.env.NEXTAUTH_SECRET,
-  session: {
-    strategy: "jwt",
-  },
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_ID ?? "",
-      clientSecret: process.env.GOOGLE_SECRET ?? "",
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      const dbUser = await prisma.user.findFirst({
-        where: { email: token.email as string },
-      });
-      if (!dbUser) {
-        if (user) {
-          token.id = user?.id;
-        }
-        return token;
-      }
-      return {
-        id: dbUser.id,
-        name: dbUser.name,
-        email: dbUser.email,
-      };
-    },
-    async session({ session, token }: any) {
-      if (session.user) {
-        session.user.id = token.id as string;
-        session.user.name = token.name as string;
-        session.user.email = token.email as string;
-      }
-      return session;
-    },
-  },
-};
-
-export const getServerSession = cache(async () => {
-  return originalGetServerSession(authOptions);
-});
+import { authOptions } from "@/src/lib/auth";
 
 const handler = NextAuth(authOptions);
